tutorial: Update startup script to check for files

+++ tutorial/setup_tutorial.sh
@@ -1,7 +1,7 @@
 set +e
 mkdir -p /tutorial/
-mv /home/snn.net /tutorial/
-mv /home/loihi.yaml /tutorial/
-mv /home/arch.yaml /tutorial/
-mv /home/challenge.py /tutorial/
-mv /home/dvs_challenge.npz /tutorial/
+[ -f /home/snn.net ] && mv /home/snn.net /tutorial/
+[ -f /home/arch.yaml ] && mv /home/arch.yaml /tutorial/
+[ -f /home/loihi.yaml ] && mv /home/loihi.yaml /tutorial/
+[ -f /home/dvs_challenge.npz ] && mv /home/dvs_challenge.npz /tutorial/
+[ -f /home/dvs_challenge.py ] && mv /home/dvs_challenge.py /tutorial/
